Add mobile WebKit main-thread render fallback
