+++ .github/workflows/playground-cross-browser.yml
@@ -13,7 +13,9 @@ on:
       - "scripts/build-web-demo.sh"
       - "scripts/build-python-worker.mjs"
       - "scripts/playground-browser-matrix-smoke.mjs"
+      - "scripts/playground-browser-runtime-gate.mjs"
       - "scripts/playground-example-browser-smoke.mjs"
+      - "scripts/playground-main-thread-mobile-smoke.mjs"
       - "web/**"
   pull_request:
     paths:
@@ -25,7 +27,9 @@ on:
       - "scripts/build-web-demo.sh"
       - "scripts/build-python-worker.mjs"
       - "scripts/playground-browser-matrix-smoke.mjs"
+      - "scripts/playground-browser-runtime-gate.mjs"
       - "scripts/playground-example-browser-smoke.mjs"
+      - "scripts/playground-main-thread-mobile-smoke.mjs"
       - "web/**"
   workflow_dispatch:
 
@@ -128,7 +132,12 @@ jobs:
           NOON_PLAYGROUND_BROWSER: ${{ matrix.browser }}
           NOON_PLAYGROUND_PROFILE: ${{ matrix.profile }}
           NOON_PLAYGROUND_MATRIX_ARTIFACTS: browser-smoke-artifacts/playground-matrix/${{ matrix.browser }}-${{ matrix.profile }}
-        run: node scripts/playground-browser-matrix-smoke.mjs
+        run: node scripts/playground-browser-runtime-gate.mjs
+      - name: Test main-thread mobile render fallback
+        if: matrix.browser == 'webkit' && matrix.profile == 'mobile-dpr2'
+        env:
+          NOON_PLAYGROUND_MATRIX_ARTIFACTS: browser-smoke-artifacts/playground-matrix/${{ matrix.browser }}-${{ matrix.profile }}
+        run: node scripts/playground-main-thread-mobile-smoke.mjs
       - name: Upload playground diagnostics
         if: always()
         uses: actions/upload-artifact@v7
